server/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
environ==1.0
idna==3.11
Pillow==11.2.1
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.5
stripe==14.1.0
typing_extensions==4.15.0
urllib3==2.6.3
